% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_sk_unif.R
\name{max_sk_unif_expm}
\alias{max_sk_unif_expm}
\title{Point-wise maximum of Skeletoid and Uniformization computed in parallel}
\usage{
max_sk_unif_expm(Q, t_pow, eps)
}
\arguments{
\item{Q}{rate matrix.}

\item{t_pow}{the t in exp(tQ)}

\item{eps}{desired error tolerance}
}
\value{
exp(tQ) up to tolerance eps in op-1 norm.
}
\description{
Point-wise maximum of Skeletoid and Uniformization computed in parallel
}
