% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_sk_unif.R
\name{max_sk_unif_vtexpm}
\alias{max_sk_unif_vtexpm}
\title{Point-wise maximum of Skeletoid and Uniformization computed in parallel}
\usage{
max_sk_unif_vtexpm(Q, t_pow, v, eps, verbose = FALSE)
}
\arguments{
\item{Q}{rate matrix.}

\item{t_pow}{the t in exp(tQ)}

\item{v}{initial distribution, must be sparse of class dgCMatrix}

\item{eps}{desired error tolerance}
}
\value{
v(t) = v %*% exp(tQ) up to tolerance eps.
}
\description{
Point-wise maximum of Skeletoid and Uniformization computed in parallel
}
