% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletoid_expm.R
\name{skeletoid_expm}
\alias{skeletoid_expm}
\title{Skeletoid}
\usage{
skeletoid_expm(Q, K, t_pow = 1, eps, n_sq = K, sparse)
}
\arguments{
\item{Q}{rate matrix.}

\item{K}{precision level. If provided, eps is ignored.}

\item{t_pow}{the t in exp(tQ)}

\item{eps}{desired error tolerance}
}
\value{
exp(tQ) up to tolerance eps in op-1 norm.
}
\description{
Skeletoid
}
