% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletoid_vtexpm.R
\name{skeletoid_vtexpm}
\alias{skeletoid_vtexpm}
\title{Skeletoid}
\usage{
skeletoid_vtexpm(
  Q,
  K,
  t_pow,
  v,
  eps,
  verbose = FALSE,
  sparse,
  n3 = glovars$N3,
  part_K
)
}
\arguments{
\item{Q}{rate matrix.}

\item{K}{precision level. If provided, eps is ignored.}

\item{t_pow}{the t in exp(tQ)}

\item{v}{initial distribution, must be sparse of class dgCMatrix}

\item{eps}{desired error tolerance}
}
\value{
v(t) = v %*% exp(tQ) up to tolerance eps.
}
\description{
Skeletoid
}
