% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_vtexpm.R
\name{unif_vtexpm}
\alias{unif_vtexpm}
\title{Uniformization}
\usage{
unif_vtexpm(Q, K, t_pow = 1, v, eps, sparse, verbose = FALSE)
}
\arguments{
\item{Q}{rate matrix.}

\item{K}{precision level. If provided, eps is ignored.}

\item{t_pow}{the t in exp(tQ)}

\item{v}{initial distribution, must be sparse of class dgCMatrix}

\item{eps}{desired error tolerance}
}
\value{
v(t) = v %*% exp(tQ) up to tolerance eps.
}
\description{
Uniformization
}
